Extract SQL invoked inlined functions tests into AbstractTestEngineOnlyQueries
