+++ presto-native-tests/src/test/java/com/facebook/presto/nativetests/TestDistributedEngineOnlyQueries.java
@@ -92,4 +92,35 @@ public class TestDistributedEngineOnlyQueries
         @Language("SQL") String sql = DateTimeFormatter.ofPattern("'SELECT TIME '''HH:mm:ss''").format(localTimeThatDidNotOccurOn20120401);
         assertQueryFails(sql, timeTypeUnsupportedError);
     }
+
+    // todo: turn on these test cases when the sql invoked functions are extracted into a plugin module.
+    @Override
+    @Test(enabled = false)
+    public void testArraySplitIntoChunks()
+    {
+    }
+
+    @Override
+    @Test(enabled = false)
+    public void testCrossJoinWithArrayNotContainsCondition()
+    {
+    }
+
+    @Override
+    @Test(enabled = false)
+    public void testSamplingJoinChain()
+    {
+    }
+
+    @Override
+    @Test(enabled = false)
+    public void testKeyBasedSampling()
+    {
+    }
+
+    @Override
+    @Test(enabled = false)
+    public void testDefaultSamplingPercent()
+    {
+    }
 }
